refactor(core): :recycle: rename branding file to get-branding
